Implementation/mechanisms

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mechanisms/Intake/Intake.java
@@ -14,7 +14,7 @@ public class Intake extends EctoMechanism {
     IntakeConfig intakeConfig;
     Motor intakeMotor;
 
-    public void turnOn(int powerPecentage){
+    public void turnOn(double powerPecentage){
         intakeMotor.set(powerPecentage);
     }
